V1 widgets folder removed
