+++ lib/v2/main.dart
@@ -23,7 +23,7 @@ import 'package:seeds/v2/screens/authentication/verification/verification_screen
 import 'package:seeds/v2/screens/onboarding/onboarding_screen.dart';
 import 'package:seeds/v2/screens/sign_up/signup_screen.dart';
 import 'package:seeds/v2/seeds_material_app.dart';
-import 'package:seeds/widgets/splash_screen.dart';
+import 'package:seeds/v2/screens/splash_screen.dart';
 
 bool get isInDebugMode {
   var inDebugMode = false;
